edit and policy

+++ app/Http/Controllers/ImagenController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Imagen;
+use App\Establecimiento;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Intervention\Image\Facades\Image;
@@ -38,24 +39,37 @@ class ImagenController extends Controller
 
     // Eliminar una imagen de la BD y del servidor
     public function destroy(Request $request)
-    {
+    {   
+
+        // Validacion
+        $uuid = $request->get('uuid');
+        $establecimiento = Establecimiento::where('uuid', $uuid)->first();
+        $this->authorize('delete', $establecimiento);
+
+        // imagen a eliminar
         $imagen = $request->get('imagen');
 
         if(File::exists('storage/' . $imagen)){  // File el archivo ->  Illuminate\Support\Facades\File;
+
+            // Elimina imagen del servidor
             File::delete('storage/' . $imagen);
 
+            // Eliminar imagen de la bd
+            Imagen::where('ruta_imagen', $imagen)->delete(); 
+
+            $respuesta = [
+                'mensaje' => 'Imagen Eliminada',
+                'imagen' => $imagen
+            ];
         }
 
         // Elimina la imagen de la BD
-        Imagen::where('ruta_imagen', '=', $imagen)->delete();
+        // Imagen::where('ruta_imagen', '=', $imagen)->delete(); // 
 
         // $imagenEliminar = Imagen::where('ruta_imagen', '=', $imagen)->firstOrFail();
         // Imagen::destroy($imagenEliminar->id);
 
-        $respuesta = [
-            'mensaje' => 'Imagen Eliminada',
-            'imagen' => $imagen
-        ];
+        
 
         return response()->json($respuesta);
     }
